+++ tests/Datasets/UnlabeledTest.php
@@ -603,7 +603,7 @@ class UnlabeledTest extends TestCase
                 'type' => 'continuous',
                 'mean' => 0.3333333333333333,
                 'variance' => 9.792222222222222,
-                'std_dev' => 3.129252661934191,
+                'stddev' => 3.129252661934191,
                 'skewness' => -0.4481030843690633,
                 'kurtosis' => -1.1330702741786107,
                 'min' => -5.0,